Added int size checks

+++ engine/src/defines.h
@@ -11,17 +11,17 @@
 
 #include <stdint.h>
 
-typedef uint8_t     u8;
-typedef uint16_t    u16;
-typedef uint32_t    u32;
-typedef uint64_t    u64;
-typedef int8_t      i8;
-typedef int16_t     i16;
-typedef int32_t     i32;
-typedef int64_t     i64;
-typedef float       f32;
-typedef double      f64;
-typedef _Bool       b8;
+typedef uint8_t                     u8;
+typedef uint16_t                    u16;
+typedef uint32_t                    u32;
+typedef unsigned long long int      u64;
+typedef int8_t                      i8;
+typedef int16_t                     i16;
+typedef int32_t                     i32;
+typedef signed long long int        i64;
+typedef float                       f32;
+typedef double                      f64;
+typedef _Bool                       b8;
 
 #define true    ((b8)1)
 #define false   ((b8)0)
